requirements.txt
# Django Core
django==5.2.5
djangorestframework==3.16.1

# Database
psycopg2-binary==2.9.10

# Configuration
python-decouple==3.8

# CORS
django-cors-headers==4.7.0

# Image Processing
Pillow==11.3.0

# HTTP Requests
requests==2.32.5

# Geographic Data
folium
geopy

# Math & Optimization (für Running Dinner Algorithmus)
numpy
scipy
PuLP
# Production Server
gunicorn

# Monitoring & Logging
sentry-sdk[django]==2.17.0

# Development tools (kommentiert für Produktion)
# django-debug-toolbar==4.4.6
# ipython==8.26.0
# django-extensions==3.2.3





